requirements.txt

fastapi

uvicorn

requests

git+https://github.com/ggerganov/llama.cpp

gradio

numpy

transformers

openai

tinydb

python-multipart

jinja2

diffusers @ git+https://github.com/huggingface/diffusers.git@77b2162817acbc846fde0342c9e81c8e267380b6
fastapi==0.112.1

huggingface-hub==0.24.6

llama_cpp_python==0.2.89

numpy==1.26.4
openai==1.42.0
pydantic==2.8.2
pydantic_core==2.20.1

safetensors==0.4.4
semantic-version==2.10.0

setuptools==74.0.0

spaces==0.29.3
starlette==0.38.2

tinydb==4.8.0
tokenizers==0.19.1

tqdm==4.66.5
transformers==4.42.4

uvicorn==0.30.6
websockets==13.0
xformers==0.0.27.post2
